CR Refactor
