--- app/controllers/editor/editor_controller.rb
@@ -1,7 +1,0 @@
-module Editor
-  class EditorController < ApplicationController
-    include EditorHelper
-
-    before_filter :editor_users_only
-  end
-end
