+++ app/views/carto/editor/visualizations/show.html.erb
